+++ monte-cover/src/montecover/did/__init__.py
@@ -1,5 +1,5 @@
 """Monte Carlo coverage simulations for DiD."""
 
-from montecover.did.did_multi import DIDMultiCoverageSimulation
+from montecover.did.did_pa_multi import DIDMultiCoverageSimulation
 
 __all__ = ["DIDMultiCoverageSimulation"]
