+++ results/did/did_multi_detailed.csv
@@ -1,25 +1,25 @@
 Learner g,Learner m,Score,In-sample-norm.,DGP,level,Coverage,CI Length,Bias,Uniform Coverage,Uniform CI Length,repetition
-Linear,Linear,experimental,False,1,0.9,0.8813333333333334,0.5961686505266492,0.15102133299318632,0.834,0.9296789699097547,1000
-Linear,Linear,experimental,False,1,0.95,0.9350833333333334,0.7103787623398975,0.15102133299318632,0.908,1.0206042134000024,1000
-Linear,Linear,experimental,False,4,0.9,0.6619166666666666,1.986651429023609,0.8637379841284726,0.52,2.882026229739649,1000
-Linear,Linear,experimental,False,4,0.95,0.7520833333333333,2.3672411860366602,0.8637379841284726,0.617,3.2128015018838694,1000
-Linear,Linear,experimental,False,6,0.9,0.902,2.00497022725443,0.4810970023674208,0.911,2.9031760219044904,1000
-Linear,Linear,experimental,False,6,0.95,0.9528333333333334,2.3890693804633027,0.4810970023674208,0.951,3.238600319213653,1000
-Linear,Linear,experimental,True,1,0.9,0.8800833333333333,0.5960959285920073,0.15123045002955704,0.833,0.928366902428871,1000
-Linear,Linear,experimental,True,1,0.95,0.9353333333333333,0.7102921088100949,0.15123045002955704,0.913,1.0197692316323934,1000
-Linear,Linear,experimental,True,4,0.9,0.6590833333333334,1.9857772419115973,0.8626296646616975,0.52,2.8815675975045956,1000
-Linear,Linear,experimental,True,4,0.95,0.7531666666666667,2.3661995278445764,0.8626296646616975,0.619,3.2112254886896605,1000
-Linear,Linear,experimental,True,6,0.9,0.9019166666666666,2.004527154284187,0.4808213402748324,0.905,2.9011439332421705,1000
-Linear,Linear,experimental,True,6,0.95,0.952,2.3885414264557423,0.4808213402748324,0.949,3.237049999080413,1000
-Linear,Linear,observational,False,1,0.9,0.90575,0.6906210074073058,0.16339132921843413,0.897,1.0727694533982786,1000
-Linear,Linear,observational,False,1,0.95,0.9518333333333334,0.8229256873110352,0.16339132921843413,0.946,1.1792381527650921,1000
-Linear,Linear,observational,False,4,0.9,0.7795,2.9571361297998364,0.9287818108191471,0.691,4.19740184270945,1000
-Linear,Linear,observational,False,4,0.95,0.8571666666666666,3.5236450324955486,0.9287818108191471,0.78,4.697698545475343,1000
-Linear,Linear,observational,False,6,0.9,0.9076666666666666,2.5544833881976365,0.6416394592068319,0.907,3.6632944671996808,1000
-Linear,Linear,observational,False,6,0.95,0.9550833333333334,3.0438546980332184,0.6416394592068319,0.958,4.0959649390037125,1000
-Linear,Linear,observational,True,1,0.9,0.8973333333333333,0.668426359641987,0.16249117038954014,0.895,1.038888800002171,1000
-Linear,Linear,observational,True,1,0.95,0.9465833333333333,0.7964791333096312,0.16249117038954014,0.939,1.1417719060066087,1000
-Linear,Linear,observational,True,4,0.9,0.7675,2.823122727518332,0.9146880810616262,0.686,4.025190417054882,1000
-Linear,Linear,observational,True,4,0.95,0.8508333333333333,3.3639582144020532,0.9146880810616262,0.768,4.499697598146944,1000
-Linear,Linear,observational,True,6,0.9,0.89775,2.4371258648326704,0.6190705910995143,0.893,3.5036401350107074,1000
-Linear,Linear,observational,True,6,0.95,0.9521666666666666,2.9040145837876374,0.6190705910995143,0.955,3.9130910221754434,1000
+Linear,Linear,experimental,False,1,0.9,0.8770833333333333,0.5956631078920439,0.15393126826790132,0.846,0.9280642120939564,1000
+Linear,Linear,experimental,False,1,0.95,0.9353333333333333,0.70977637113606,0.15393126826790132,0.907,1.01868658603788,1000
+Linear,Linear,experimental,False,4,0.9,0.6671666666666666,1.9827195149925345,0.8550657449695865,0.505,2.8772921242320737,1000
+Linear,Linear,experimental,False,4,0.95,0.758,2.362556021493784,0.8550657449695865,0.628,3.2083215653145722,1000
+Linear,Linear,experimental,False,6,0.9,0.9025,2.00173937868972,0.49027975353015657,0.903,2.8952108652661397,1000
+Linear,Linear,experimental,False,6,0.95,0.9499166666666666,2.385219586948198,0.49027975353015657,0.939,3.230482678959989,1000
+Linear,Linear,experimental,True,1,0.9,0.8771666666666667,0.5956064011739037,0.15426338756081115,0.842,0.9278381417397478,1000
+Linear,Linear,experimental,True,1,0.95,0.9339166666666666,0.7097088009137527,0.15426338756081115,0.912,1.0187258089279716,1000
+Linear,Linear,experimental,True,4,0.9,0.6675,1.9828114892881032,0.8559251553294747,0.507,2.8761724678009353,1000
+Linear,Linear,experimental,True,4,0.95,0.7571666666666667,2.362665615626578,0.8559251553294747,0.618,3.204786768191425,1000
+Linear,Linear,experimental,True,6,0.9,0.9008333333333334,2.002017704071254,0.4900985674785569,0.904,2.897432562708396,1000
+Linear,Linear,experimental,True,6,0.95,0.9494166666666666,2.3855512320957373,0.4900985674785569,0.945,3.231599891589011,1000
+Linear,Linear,observational,False,1,0.9,0.9048333333333334,0.6855202939670088,0.16429464571110677,0.905,1.0652155650078814,1000
+Linear,Linear,observational,False,1,0.95,0.9518333333333334,0.8168478123715063,0.16429464571110677,0.951,1.1708094409885172,1000
+Linear,Linear,observational,False,4,0.9,0.7851666666666667,2.7948549075952385,0.8941603312052668,0.689,4.003170404405238,1000
+Linear,Linear,observational,False,4,0.95,0.8580833333333334,3.3302750294286816,0.8941603312052668,0.786,4.4764083966934205,1000
+Linear,Linear,observational,False,6,0.9,0.8918333333333334,2.540792582093424,0.6659145487088738,0.893,3.64735417288717,1000
+Linear,Linear,observational,False,6,0.95,0.9475,3.0275410963583322,0.6659145487088738,0.941,4.076810932574564,1000
+Linear,Linear,observational,True,1,0.9,0.8975,0.6634200132755311,0.16175894761877385,0.888,1.0319404210431198,1000
+Linear,Linear,observational,True,1,0.95,0.94725,0.790513703674064,0.16175894761877385,0.94,1.1332864016074393,1000
+Linear,Linear,observational,True,4,0.9,0.7758333333333334,2.7290002317619733,0.8973641897321528,0.674,3.912420538282281,1000
+Linear,Linear,observational,True,4,0.95,0.8505833333333334,3.2518043432035615,0.8973641897321528,0.792,4.369245108445732,1000
+Linear,Linear,observational,True,6,0.9,0.8846666666666666,2.4411244723625476,0.6398322462292214,0.892,3.513044851639997,1000
+Linear,Linear,observational,True,6,0.95,0.9423333333333334,2.9087792185359564,0.6398322462292214,0.945,3.923376128802193,1000
