+++ results/did/did_multi_eventstudy.csv
@@ -1,25 +1,25 @@
 Learner g,Learner m,Score,In-sample-norm.,DGP,level,Coverage,CI Length,Bias,Uniform Coverage,Uniform CI Length,repetition
-Linear,Linear,experimental,False,1,0.9,0.8746666666666666,0.42536022351351016,0.11129964461197367,0.856,0.6065601425726624,1000
-Linear,Linear,experimental,False,1,0.95,0.9343333333333333,0.5068479680392762,0.11129964461197367,0.923,0.6740692167988962,1000
-Linear,Linear,experimental,False,4,0.9,0.6018333333333333,1.9789774969843743,0.9612815234569009,0.504,2.557725621904857,1000
-Linear,Linear,experimental,False,4,0.95,0.6915,2.3580971320185626,0.9612815234569009,0.606,2.9077454096108863,1000
-Linear,Linear,experimental,False,6,0.9,0.9026666666666666,2.0021783896522174,0.475645152957695,0.913,2.5813810385024185,1000
-Linear,Linear,experimental,False,6,0.95,0.9546666666666667,2.3857427007749936,0.475645152957695,0.954,2.939126540032628,1000
-Linear,Linear,experimental,True,1,0.9,0.8736666666666666,0.4252780101030476,0.11135661295209692,0.855,0.6067206852239997,1000
-Linear,Linear,experimental,True,1,0.95,0.9338333333333334,0.50675000471846,0.11135661295209692,0.919,0.6737727543313963,1000
-Linear,Linear,experimental,True,4,0.9,0.604,1.9778347169116364,0.9610017220191109,0.506,2.5570866190887442,1000
-Linear,Linear,experimental,True,4,0.95,0.69,2.356735425573614,0.9610017220191109,0.614,2.906700104748762,1000
-Linear,Linear,experimental,True,6,0.9,0.9028333333333334,2.001368819203873,0.47488976645105485,0.91,2.5788107797213202,1000
-Linear,Linear,experimental,True,6,0.95,0.9543333333333334,2.3847780380866523,0.47488976645105485,0.954,2.9346744455419307,1000
-Linear,Linear,observational,False,1,0.9,0.9048333333333334,0.4936490463630555,0.11703447731264664,0.911,0.7026803919052239,1000
-Linear,Linear,observational,False,1,0.95,0.9523333333333334,0.5882191193312043,0.11703447731264664,0.953,0.7817258159201687,1000
-Linear,Linear,observational,False,4,0.9,0.7483333333333334,3.110852841424021,1.0162956359155522,0.691,3.927457336373739,1000
-Linear,Linear,observational,False,4,0.95,0.8321666666666666,3.7068097917596994,1.0162956359155522,0.775,4.489588602429944,1000
-Linear,Linear,observational,False,6,0.9,0.908,2.6065175455108682,0.657901837298179,0.912,3.319528013865773,1000
-Linear,Linear,observational,False,6,0.95,0.9611666666666666,3.1058572207068265,0.657901837298179,0.96,3.78925559700768,1000
-Linear,Linear,observational,True,1,0.9,0.8983333333333333,0.478045989951027,0.1165385646041097,0.898,0.6810014878305387,1000
-Linear,Linear,observational,True,1,0.95,0.9476666666666667,0.5696269308742887,0.1165385646041097,0.951,0.7570193996194396,1000
-Linear,Linear,observational,True,4,0.9,0.7318333333333333,2.961887231450272,1.0118391967276548,0.664,3.755668505865932,1000
-Linear,Linear,observational,True,4,0.95,0.8186666666666667,3.5293063192928438,1.0118391967276548,0.771,4.287491165909032,1000
-Linear,Linear,observational,True,6,0.9,0.8983333333333333,2.4694251965466845,0.6253200792538618,0.908,3.1600238226116533,1000
-Linear,Linear,observational,True,6,0.95,0.9541666666666666,2.9425016113546496,0.6253200792538618,0.955,3.596198019745218,1000
+Linear,Linear,experimental,False,1,0.9,0.8676666666666666,0.42543423226976995,0.11378217794214893,0.82,0.607026998501089,1000
+Linear,Linear,experimental,False,1,0.95,0.921,0.5069361549116113,0.11378217794214893,0.893,0.6744173882852083,1000
+Linear,Linear,experimental,False,4,0.9,0.6021666666666666,1.9754395142835157,0.9518163003680035,0.488,2.5538078783006415,1000
+Linear,Linear,experimental,False,4,0.95,0.7053333333333334,2.353881365607505,0.9518163003680035,0.602,2.9021513463156756,1000
+Linear,Linear,experimental,False,6,0.9,0.8995,2.001327059411559,0.49627387221768626,0.899,2.5768571826106514,1000
+Linear,Linear,experimental,False,6,0.95,0.949,2.3847282782249866,0.49627387221768626,0.942,2.9364211171912067,1000
+Linear,Linear,experimental,True,1,0.9,0.8666666666666666,0.42531823168334737,0.1141016026224358,0.814,0.6067463378989032,1000
+Linear,Linear,experimental,True,1,0.95,0.9231666666666666,0.5067979316874602,0.1141016026224358,0.889,0.6733443239100646,1000
+Linear,Linear,experimental,True,4,0.9,0.5973333333333334,1.975365775525183,0.9534736795550611,0.491,2.5532803279562994,1000
+Linear,Linear,experimental,True,4,0.95,0.7055,2.353793500457543,0.9534736795550611,0.61,2.905573717129059,1000
+Linear,Linear,experimental,True,6,0.9,0.8978333333333334,2.0019957931284336,0.4959848232257912,0.9,2.579876093556324,1000
+Linear,Linear,experimental,True,6,0.95,0.9478333333333334,2.385525123597028,0.4959848232257912,0.944,2.9374235003380655,1000
+Linear,Linear,observational,False,1,0.9,0.9013333333333333,0.4873576971767433,0.11751015177996685,0.898,0.693580479033222,1000
+Linear,Linear,observational,False,1,0.95,0.9505,0.5807225143948788,0.11751015177996685,0.944,0.7712324745986096,1000
+Linear,Linear,observational,False,4,0.9,0.762,2.9253154914773845,0.972649026341708,0.692,3.7241194943689173,1000
+Linear,Linear,observational,False,4,0.95,0.8396666666666667,3.485728403286012,0.972649026341708,0.787,4.249241903518029,1000
+Linear,Linear,observational,False,6,0.9,0.8926666666666666,2.596496390444534,0.6912487923777589,0.89,3.308229651637506,1000
+Linear,Linear,observational,False,6,0.95,0.9455,3.0939162779434826,0.6912487923777589,0.948,3.7769009967220812,1000
+Linear,Linear,observational,True,1,0.9,0.8956666666666666,0.4727299379079462,0.11588912406999206,0.894,0.6735775257798401,1000
+Linear,Linear,observational,True,1,0.95,0.944,0.5632924641632966,0.11588912406999206,0.938,0.7483724833065163,1000
+Linear,Linear,observational,True,4,0.9,0.7468333333333333,2.8437612590899457,0.9776121628787491,0.682,3.627947639783714,1000
+Linear,Linear,observational,True,4,0.95,0.8328333333333334,3.3885505415923616,0.9776121628787491,0.775,4.133831167481912,1000
+Linear,Linear,observational,True,6,0.9,0.8851666666666667,2.4823044000101313,0.660809427159733,0.894,3.171860461977204,1000
+Linear,Linear,observational,True,6,0.95,0.9428333333333334,2.957848128834568,0.660809427159733,0.945,3.6158624454294848,1000
