+++ results/did/did_multi_group.csv
@@ -1,25 +1,25 @@
 Learner g,Learner m,Score,In-sample-norm.,DGP,level,Coverage,CI Length,Bias,Uniform Coverage,Uniform CI Length,repetition
-Linear,Linear,experimental,False,1,0.9,0.876,0.5329948975220807,0.13579236732447947,0.846,0.683878514526821,1000
-Linear,Linear,experimental,False,1,0.95,0.9296666666666666,0.6351025926987942,0.13579236732447947,0.917,0.7732276242829601,1000
-Linear,Linear,experimental,False,4,0.9,0.658,2.1938722973937375,0.9674173487082218,0.525,2.767821034152252,1000
-Linear,Linear,experimental,False,4,0.95,0.746,2.614160080335666,0.9674173487082218,0.619,3.1440243876915375,1000
-Linear,Linear,experimental,False,6,0.9,0.9006666666666666,2.211125648458694,0.5338317649415134,0.919,2.783739324710087,1000
-Linear,Linear,experimental,False,6,0.95,0.9543333333333334,2.6347187161594587,0.5338317649415134,0.957,3.1662952947465888,1000
-Linear,Linear,experimental,True,1,0.9,0.8766666666666666,0.5328699306003187,0.1358281841054489,0.848,0.6847832796423149,1000
-Linear,Linear,experimental,True,1,0.95,0.9263333333333333,0.6349536854271081,0.1358281841054489,0.911,0.7727182150404185,1000
-Linear,Linear,experimental,True,4,0.9,0.6586666666666666,2.1921756345553276,0.9652578245005206,0.525,2.7628627454822605,1000
-Linear,Linear,experimental,True,4,0.95,0.7473333333333334,2.6121383818679704,0.9652578245005206,0.624,3.144606886076096,1000
-Linear,Linear,experimental,True,6,0.9,0.8993333333333333,2.2104603719936415,0.5344678450501394,0.912,2.784794352421448,1000
-Linear,Linear,experimental,True,6,0.95,0.954,2.6339259903570538,0.5344678450501394,0.957,3.164126890302782,1000
-Linear,Linear,observational,False,1,0.9,0.9026666666666666,0.6123404760096205,0.1438612793224918,0.904,0.7860078542582829,1000
-Linear,Linear,observational,False,1,0.95,0.9546666666666667,0.7296486809463548,0.1438612793224918,0.952,0.8872672286776676,1000
-Linear,Linear,observational,False,4,0.9,0.781,3.3074883394733083,1.0352240102334336,0.719,4.086258476471432,1000
-Linear,Linear,observational,False,4,0.95,0.8563333333333334,3.9411154393527843,1.0352240102334336,0.813,4.663160227096469,1000
-Linear,Linear,observational,False,6,0.9,0.899,2.789010379110977,0.6998566663431153,0.912,3.4660289192044,1000
-Linear,Linear,observational,False,6,0.95,0.9553333333333334,3.323310844197804,0.6998566663431153,0.96,3.9528394153130857,1000
-Linear,Linear,observational,True,1,0.9,0.9023333333333333,0.5949231101949434,0.1440501200808088,0.897,0.763308057016272,1000
-Linear,Linear,observational,True,1,0.95,0.9486666666666667,0.7088946094940543,0.1440501200808088,0.936,0.8623401634425215,1000
-Linear,Linear,observational,True,4,0.9,0.7643333333333334,3.142810509409861,1.0233709921822023,0.706,3.8984668952364427,1000
-Linear,Linear,observational,True,4,0.95,0.8463333333333334,3.7448897018841176,1.0233709921822023,0.804,4.447829548666937,1000
-Linear,Linear,observational,True,6,0.9,0.8976666666666666,2.6701046408964046,0.6759660173055291,0.902,3.329342240927625,1000
-Linear,Linear,observational,True,6,0.95,0.9526666666666667,3.181625918173329,0.6759660173055291,0.953,3.7967419653105576,1000
+Linear,Linear,experimental,False,1,0.9,0.865,0.5332689974285741,0.14125883992697605,0.84,0.684967458838645,1000
+Linear,Linear,experimental,False,1,0.95,0.923,0.6354292028822721,0.14125883992697605,0.898,0.7737658489393863,1000
+Linear,Linear,experimental,False,4,0.9,0.6613333333333332,2.187885379540459,0.9604906586192218,0.498,2.7575340571302047,1000
+Linear,Linear,experimental,False,4,0.95,0.7556666666666666,2.607026227706741,0.9604906586192218,0.602,3.1362004034484614,1000
+Linear,Linear,experimental,False,6,0.9,0.8976666666666666,2.2072703099683895,0.5422311709957129,0.907,2.7789385342819997,1000
+Linear,Linear,experimental,False,6,0.95,0.9506666666666667,2.6301247970013075,0.5422311709957129,0.941,3.1616330454003854,1000
+Linear,Linear,experimental,True,1,0.9,0.8643333333333334,0.5331656474907569,0.1416028173409448,0.837,0.6848091387598605,1000
+Linear,Linear,experimental,True,1,0.95,0.9223333333333333,0.6353060538356899,0.1416028173409448,0.904,0.7729755854524409,1000
+Linear,Linear,experimental,True,4,0.9,0.6636666666666666,2.1871578881758738,0.9618357917015027,0.495,2.757365097655949,1000
+Linear,Linear,experimental,True,4,0.95,0.7556666666666666,2.606159368279076,0.9618357917015027,0.604,3.1339986032881795,1000
+Linear,Linear,experimental,True,6,0.9,0.8983333333333333,2.2070772542069657,0.5424515866558068,0.899,2.779655081198074,1000
+Linear,Linear,experimental,True,6,0.95,0.9513333333333334,2.629894756872995,0.5424515866558068,0.942,3.16003069657915,1000
+Linear,Linear,observational,False,1,0.9,0.902,0.6106049709052258,0.14629034271149396,0.89,0.7831902414227334,1000
+Linear,Linear,observational,False,1,0.95,0.947,0.727580699064038,0.14629034271149396,0.942,0.8844844476393456,1000
+Linear,Linear,observational,False,4,0.9,0.7813333333333333,3.0907797972605278,0.9943909423280353,0.706,3.8523704910840313,1000
+Linear,Linear,observational,False,4,0.95,0.8523333333333334,3.682891284376495,0.9943909423280353,0.809,4.39501177360628,1000
+Linear,Linear,observational,False,6,0.9,0.89,2.7865039885472247,0.7301188763670151,0.894,3.468312373428792,1000
+Linear,Linear,observational,False,6,0.95,0.95,3.3203242956346655,0.7301188763670151,0.945,3.9541003782057595,1000
+Linear,Linear,observational,True,1,0.9,0.8946666666666666,0.5896388541244254,0.14371827951522767,0.878,0.7568642372370822,1000
+Linear,Linear,observational,True,1,0.95,0.9423333333333334,0.7025980300212064,0.14371827951522767,0.936,0.8549636526386375,1000
+Linear,Linear,observational,True,4,0.9,0.777,3.0325048545897997,1.0070643044350587,0.698,3.7834488519755434,1000
+Linear,Linear,observational,True,4,0.95,0.8456666666666667,3.6134524072847696,1.0070643044350587,0.784,4.309691278217546,1000
+Linear,Linear,observational,True,6,0.9,0.8863333333333334,2.674147344485724,0.6969076659422232,0.89,3.3341905763684103,1000
+Linear,Linear,observational,True,6,0.95,0.9453333333333334,3.1864430966172987,0.6969076659422232,0.943,3.804017709516315,1000
