+++ results/did/did_multi_time.csv
@@ -1,25 +1,25 @@
 Learner g,Learner m,Score,In-sample-norm.,DGP,level,Coverage,CI Length,Bias,Uniform Coverage,Uniform CI Length,repetition
-Linear,Linear,experimental,False,1,0.9,0.862,0.4928810792988786,0.13124379587670076,0.839,0.6315410055040674,1000
-Linear,Linear,experimental,False,1,0.95,0.9253333333333333,0.5873040301327268,0.13124379587670076,0.905,0.7136054232732442,1000
-Linear,Linear,experimental,False,4,0.9,0.5016666666666667,1.9657848813119838,1.0502850008012623,0.473,2.259694817907404,1000
-Linear,Linear,experimental,False,4,0.95,0.5956666666666667,2.34237716085755,1.0502850008012623,0.576,2.6247319895995456,1000
-Linear,Linear,experimental,False,6,0.9,0.915,1.9600200144024984,0.46223102289234697,0.913,2.2566500717930382,1000
-Linear,Linear,experimental,False,6,0.95,0.9573333333333334,2.3355078982477218,0.46223102289234697,0.955,2.6243904341111817,1000
-Linear,Linear,experimental,True,1,0.9,0.86,0.49279866312558346,0.13118024657417612,0.835,0.6314273700892165,1000
-Linear,Linear,experimental,True,1,0.95,0.9256666666666666,0.5872058252050936,0.13118024657417612,0.903,0.7135421188498914,1000
-Linear,Linear,experimental,True,4,0.9,0.4956666666666667,1.9645783402752341,1.048403876833638,0.482,2.257895538601638,1000
-Linear,Linear,experimental,True,4,0.95,0.6013333333333334,2.340939478537888,1.048403876833638,0.576,2.623735792365416,1000
-Linear,Linear,experimental,True,6,0.9,0.916,1.9596368430072597,0.4616670580420602,0.917,2.258846499216311,1000
-Linear,Linear,experimental,True,6,0.95,0.9583333333333334,2.335051321369227,0.4616670580420602,0.957,2.619661989243463,1000
-Linear,Linear,observational,False,1,0.9,0.9046666666666666,0.6029754636773904,0.14420577130973078,0.91,0.7706892509169176,1000
-Linear,Linear,observational,False,1,0.95,0.9546666666666667,0.7184895804737107,0.14420577130973078,0.954,0.8719065946335489,1000
-Linear,Linear,observational,False,4,0.9,0.682,3.3042482077466073,1.1219461153182708,0.651,3.7179171164272695,1000
-Linear,Linear,observational,False,4,0.95,0.7726666666666666,3.937254584267904,1.1219461153182708,0.738,4.337439826876367,1000
-Linear,Linear,observational,False,6,0.9,0.914,2.5295434432500326,0.631825915195643,0.917,2.9157792155421123,1000
-Linear,Linear,observational,False,6,0.95,0.9626666666666667,3.014136920674323,0.631825915195643,0.967,3.386519653775383,1000
-Linear,Linear,observational,True,1,0.9,0.8916666666666666,0.5764292428073284,0.14270494869961284,0.893,0.7373182480398929,1000
-Linear,Linear,observational,True,1,0.95,0.9456666666666667,0.6868578073004362,0.14270494869961284,0.943,0.8337504767047077,1000
-Linear,Linear,observational,True,4,0.9,0.6606666666666666,3.110533543222582,1.1123478588300635,0.623,3.5061830916939076,1000
-Linear,Linear,observational,True,4,0.95,0.7673333333333334,3.7064293244859585,1.1123478588300635,0.727,4.084196858643447,1000
-Linear,Linear,observational,True,6,0.9,0.9073333333333333,2.4168915264549797,0.611462867914882,0.895,2.7838277996862724,1000
-Linear,Linear,observational,True,6,0.95,0.9533333333333334,2.879903882493947,0.611462867914882,0.953,3.2317290279843696,1000
+Linear,Linear,experimental,False,1,0.9,0.8453333333333334,0.49324862390769003,0.1376326316621036,0.822,0.6316419735152038,1000
+Linear,Linear,experimental,False,1,0.95,0.912,0.5877419865467073,0.1376326316621036,0.891,0.7143067511536322,1000
+Linear,Linear,experimental,False,4,0.9,0.4776666666666667,1.9593068261105524,1.0466452508829949,0.45,2.2507550996198735,1000
+Linear,Linear,experimental,False,4,0.95,0.579,2.3346580819823073,1.0466452508829949,0.556,2.6163918210317183,1000
+Linear,Linear,experimental,False,6,0.9,0.8956666666666666,1.9573896704373563,0.48930237462467674,0.896,2.2565806380862594,1000
+Linear,Linear,experimental,False,6,0.95,0.9463333333333334,2.3323736500969083,0.48930237462467674,0.941,2.62032018987877,1000
+Linear,Linear,experimental,True,1,0.9,0.8463333333333334,0.4931694609078208,0.1378285637524634,0.824,0.6319264464558539,1000
+Linear,Linear,experimental,True,1,0.95,0.9093333333333333,0.587647658014302,0.1378285637524634,0.895,0.7142050985707156,1000
+Linear,Linear,experimental,True,4,0.9,0.473,1.9591517453060092,1.0480526529722811,0.45,2.250777097405993,1000
+Linear,Linear,experimental,True,4,0.95,0.5806666666666667,2.334473291806076,1.0480526529722811,0.55,2.6128556265062666,1000
+Linear,Linear,experimental,True,6,0.9,0.8966666666666666,1.9576188610509668,0.49183824648546365,0.893,2.253956782621074,1000
+Linear,Linear,experimental,True,6,0.95,0.945,2.3326467475573214,0.49183824648546365,0.943,2.6202784765893865,1000
+Linear,Linear,observational,False,1,0.9,0.8953333333333334,0.5968613600047712,0.14788675870016774,0.878,0.7631035273143418,1000
+Linear,Linear,observational,False,1,0.95,0.94,0.7112041765935568,0.14788675870016774,0.945,0.8624624885034761,1000
+Linear,Linear,observational,False,4,0.9,0.6876666666666666,3.0580400071087412,1.0638406360538493,0.653,3.462223480766848,1000
+Linear,Linear,observational,False,4,0.95,0.7736666666666666,3.643879418209516,1.0638406360538493,0.75,4.0319351473266085,1000
+Linear,Linear,observational,False,6,0.9,0.8823333333333334,2.548375931783461,0.6811130775533906,0.871,2.9334540834011715,1000
+Linear,Linear,observational,False,6,0.95,0.9406666666666667,3.036577214850038,0.6811130775533906,0.945,3.4079635812732043,1000
+Linear,Linear,observational,True,1,0.9,0.8806666666666666,0.5725784741025863,0.14380420604106553,0.867,0.7329506080582452,1000
+Linear,Linear,observational,True,1,0.95,0.9413333333333334,0.682269333377637,0.14380420604106553,0.926,0.8286474263809458,1000
+Linear,Linear,observational,True,4,0.9,0.666,2.97181956637082,1.075066363572111,0.635,3.3631545325412926,1000
+Linear,Linear,observational,True,4,0.95,0.7603333333333334,3.5411414263246734,1.075066363572111,0.748,3.9204031440435947,1000
+Linear,Linear,observational,True,6,0.9,0.8823333333333334,2.4376548170067376,0.6493921794587046,0.877,2.808933035095803,1000
+Linear,Linear,observational,True,6,0.95,0.939,2.904644869178221,0.6493921794587046,0.937,3.2552833144907303,1000
