--- results/did/multi_config.yml
@@ -1,51 +1,0 @@
-confidence_parameters:
-  level:
-  - 0.95
-  - 0.9
-dgp_parameters:
-  DGP:
-  - 1
-  - 4
-  - 6
-  n_obs:
-  - 500
-dml_parameters:
-  in_sample_normalization:
-  - true
-  - false
-  learners:
-  - ml_g: !!python/tuple
-    - Linear
-    - !!python/object:sklearn.linear_model._base.LinearRegression
-      _sklearn_version: 1.5.2
-      copy_X: true
-      fit_intercept: true
-      n_jobs: null
-      positive: false
-    ml_m: !!python/tuple
-    - Linear
-    - !!python/object:sklearn.linear_model._logistic.LogisticRegression
-      C: 1.0
-      _sklearn_version: 1.5.2
-      class_weight: null
-      dual: false
-      fit_intercept: true
-      intercept_scaling: 1
-      l1_ratio: null
-      max_iter: 100
-      multi_class: deprecated
-      n_jobs: null
-      penalty: l2
-      random_state: null
-      solver: lbfgs
-      tol: 0.0001
-      verbose: 0
-      warm_start: false
-  score:
-  - observational
-  - experimental
-simulation_parameters:
-  max_runtime: 19800
-  n_jobs: -2
-  random_seed: 42
-  repetitions: 1000
